% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/randomgames.R
\name{myfun}
\alias{myfun}
\title{random games}
=======
% Please edit documentation in R/lee.R
\name{myfun}
\alias{myfun}
\title{Very Useful function}
>>>>>>> gd
\usage{
myfun(x)
}
\arguments{
<<<<<<< HEAD
\item{x}{vari}
=======
\item{x}{any number}
>>>>>>> gd
}
\value{
\code{x + 1}
}
\description{
<<<<<<< HEAD
simple ex
}
\examples{
=======
This function simply add 1.
}
\examples{
myfun(1)
>>>>>>> gd

}
